modif 2 InputField
